+++ lib/models/ny_revenuecat_slate_config.dart
@@ -1,0 +1,10 @@
+/// This class is used to store the configuration of the RevenueCat Slate SDK.
+class NyRevenueCatSlateConfig {
+  String? androidAppId;
+  String? appleAppId;
+
+  NyRevenueCatSlateConfig({
+    required this.appleAppId,
+    required this.androidAppId,
+  });
+}