+++ lib/stubs/firebase/firebase_provider_stub.dart
@@ -1,0 +1,30 @@
+String stubFirebaseProvider() => '''
+import 'package:firebase_auth/firebase_auth.dart';
+import 'package:firebase_core/firebase_core.dart';
+import 'package:nylo_framework/nylo_framework.dart';
+import '/firebase_options.dart';
+
+class FirebaseProvider implements NyProvider {
+  @override
+  boot(Nylo nylo) async {
+
+    // firebase
+    await Firebase.initializeApp(
+      options: DefaultFirebaseOptions.currentPlatform,
+    );
+    
+    FirebaseAuth.instance.authStateChanges().listen((User? user) {
+      if (user == null) {
+        routeToInitial();
+      }
+    });
+
+    return null;
+  }
+
+  @override
+  afterBoot(Nylo nylo) async {
+    
+  }
+}
+''';